SOLR-15288: Fix test resource leak issue
