+++ solr/solrj/src/test/org/apache/solr/client/solrj/impl/CloudSolrClientTest.java
@@ -53,7 +53,6 @@ import org.apache.solr.client.solrj.response.RequestStatusState;
 import org.apache.solr.client.solrj.response.SolrPingResponse;
 import org.apache.solr.client.solrj.response.UpdateResponse;
 import org.apache.solr.cloud.AbstractDistribZkTestBase;
-import org.apache.solr.cloud.MiniSolrCloudCluster;
 import org.apache.solr.cloud.SolrCloudTestCase;
 import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.SolrDocumentList;
@@ -1121,63 +1120,4 @@ public class CloudSolrClientTest extends SolrCloudTestCase {
     assertEquals(4, prs.states.size());
   }
 
-  public void testPRSRestart() throws Exception {
-    String testCollection = "prs_restart_test";
-    MiniSolrCloudCluster cluster1 =
-        configureCluster(1)
-            .addConfig("conf1", getFile("solrj").toPath().resolve("solr").resolve("configsets").resolve("streaming").resolve("conf"))
-            .withJettyConfig(jetty -> jetty.enableV2(true))
-            .configure();
-    try {
-      CollectionAdminRequest.createCollection(testCollection, "conf1", 1, 1)
-          .setPerReplicaState(Boolean.TRUE)
-          .process(cluster1.getSolrClient());
-      cluster1.waitForActiveCollection(testCollection, 1, 1);
-
-      DocCollection c = cluster1.getSolrClient().getZkStateReader().getCollection(testCollection);
-      c.forEachReplica((s, replica) -> assertNotNull(replica.getReplicaState()));
-      String collectionPath = ZkStateReader.getCollectionPath(testCollection);
-      PerReplicaStates prs = PerReplicaStates.fetch(collectionPath, cluster.getZkClient(), null);
-      assertEquals(1, prs.states.size());
-
-      JettySolrRunner jsr = cluster1.startJettySolrRunner();
-      assertEquals(2,cluster1.getJettySolrRunners().size());
-
-      // Now let's do an add replica
-      CollectionAdminRequest
-          .addReplicaToShard(testCollection, "shard1")
-          .process(cluster1.getSolrClient());
-      cluster1.waitForActiveCollection(testCollection, 1, 2);
-      prs = PerReplicaStates.fetch(collectionPath, cluster.getZkClient(), null);
-      assertEquals(2, prs.states.size());
-      c = cluster1.getSolrClient().getZkStateReader().getCollection(testCollection);
-      prs.states.forEachEntry((s, state) -> assertEquals(Replica.State.ACTIVE, state.state));
-
-      String replicaName = null;
-      for (Replica r : c.getSlice("shard1").getReplicas()) {
-        if(r.getNodeName() .equals(jsr.getNodeName())) {
-          replicaName = r.getName();
-        }
-      }
-
-      if(replicaName != null) {
-        log.info("restarting the node : {}, state.json v: {} downreplica :{}", jsr.getNodeName(), c.getZNodeVersion(), replicaName);
-        jsr.stop();
-        c = cluster1.getSolrClient().getZkStateReader().getCollection(testCollection);
-        log.info("after down node, state.json v: {}", c.getZNodeVersion());
-        prs =  PerReplicaStates.fetch(collectionPath, cluster.getZkClient(), null);
-        PerReplicaStates.State st = prs.get(replicaName);
-        assertNotEquals(Replica.State.ACTIVE, st.state);
-        jsr.start();
-        cluster1.waitForActiveCollection(testCollection, 1, 2);
-        prs =  PerReplicaStates.fetch(collectionPath, cluster.getZkClient(), null);
-        prs.states.forEachEntry((s, state) -> assertEquals(Replica.State.ACTIVE, state.state));
-      }
-
-    } finally {
-      cluster1.shutdown();
-    log.info("SHUTDOWN CLUSTER");
-    }
-
-  }
 }
